DCAC-68: Separate out message production into factory component.
